refacto controller
