+++ src/code-runner/code-runner.module.ts
@@ -3,6 +3,7 @@ import { CodeRunnerService } from './code-runner.service';
 import { DockerManagerModule } from './docker-manager/docker-manager.module';
 import { FileManagerModule } from './file-manager/file-manager.module';
 import { ExecutionScriptGenerateManagerModule } from './execution-script-generate-manager/execution-script-generate-manager.module';
+import { CodeRunnerController } from './code-runner.controller';
 
 @Module({
   imports: [
@@ -12,5 +13,6 @@ import { ExecutionScriptGenerateManagerModule } from './execution-script-generat
   ],
   providers: [CodeRunnerService],
   exports: [CodeRunnerService],
+  controllers: [CodeRunnerController],
 })
 export class CodeRunnerModule {}
