+++ src/posts/posts.module.ts
@@ -4,10 +4,12 @@ import { UsersService } from '../users/users.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Post } from './post.entity';
 import { User } from '../users/user.entity';
+import { PostsController } from './posts.controller';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Post, User])],
   providers: [PostsService, UsersService],
   exports: [PostsService],
+  controllers: [PostsController],
 })
 export class PostsModule {}
